+++ src/App.js
@@ -155,6 +155,7 @@ import {LotteryGame} from './containers/Gamification';
 import ModalConfirm from './components/ModalConfirm';
 import ProductStamps from './containers/ProductStamps';
 import ModalComboLocation from './components/ModalComboLocation';
+import APIHandler from './network/APIHandler';
 
 /**
  * Not allow font scaling
@@ -442,9 +443,18 @@ class App extends Component {
     );
   }
 
+  async updateAppVersionsInfo(codePushVersion) {
+    const data = {
+      code_push_version: codePushVersion,
+      tag_version: appConfig.tagVersion,
+    };
+    const response = await APIHandler.user_device(data);
+  }
+
   codePushGetMetaData() {
     codePush.getUpdateMetadata().then((localPackage) => {
       store.setCodePushMetaData(localPackage);
+      this.updateAppVersionsInfo(localPackage?.label);
     });
   }
 
@@ -1679,7 +1689,10 @@ class RootRouter extends Component {
               <Stack key={appConfig.routes.modalInput} component={ModalInput} />
 
               {/* ================ MODAL CONFIRM ================ */}
-              <Stack key={appConfig.routes.modalConfirm} component={ModalConfirm} />
+              <Stack
+                key={appConfig.routes.modalConfirm}
+                component={ModalConfirm}
+              />
 
               {/* ================ COUNTRY PICKER ================ */}
               <Stack